experiments/peanutgallery/pyproject.toml
[project]
name = "peanutgallery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.12.0",
    "flask>=3.1.2",
    "nltk>=3.9.2",
    "numpy>=2.3.4",
    "ollama>=0.6.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "tqdm>=4.67.1",
]
